+++ ptai-cli-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/CliAstJob.java
@@ -24,6 +24,7 @@ public class CliAstJob extends AstJob {
     protected Path input;
     protected String includes;
     protected String excludes;
+    protected boolean useDefaultExcludes;
     protected Path output;
 
     protected BaseCommand.Reporting reporting;