+++ ptai-cli-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/commands/JsonAst.java
@@ -63,6 +63,11 @@ public class JsonAst extends BaseCommand implements Callable<Integer> {
             description = "Comma-separated list of files to exclude from scan. The syntax is the same as for includes")
     protected String excludes = null;
 
+    @CommandLine.Option(
+            names = {"--use-default-excludes"}, order = 9,
+            description = "Use default excludes list")
+    private boolean useDefaultExcludes = false;
+
     @CommandLine.ArgGroup(exclusive = true)
     BaseCommand.Reporting reports;
 
@@ -95,6 +100,7 @@ public class JsonAst extends BaseCommand implements Callable<Integer> {
                 .failIfFailed(failIfFailed).failIfUnstable(failIfUnstable)
                 .input(input).output(output)
                 .includes(includes).excludes(excludes)
+                .useDefaultExcludes(useDefaultExcludes)
                 .reporting(reports)
                 .truststore(truststore)
                 .settings(jsonSettings)